# Remove network from EDS metadata