@howardjohn 2021-06-30
merged into https://github.com/istio/istio/pull/33715, so I will close this. If we don't merge that for some reason we should reopen